Add images to paint

+++ app/src/main/java/com/example/paintingjournal/model/PaintImageMappingTable.kt
@@ -26,7 +26,7 @@ import androidx.room.ForeignKey
 )
 data class PaintImageMappingTable (
     @ColumnInfo(name = "paintIdRef")
-    val paintIdRef: Int,
+    val paintIdRef: Long,
     @ColumnInfo(index = true, name = "imageIdRef")
-    val imageIdRef: Int
+    val imageIdRef: Long
 )
